Speed up import optuna
