+++ optuna/_imports.py
@@ -1,4 +1,7 @@
+import importlib
+import types
 from types import TracebackType
+from typing import Any
 from typing import Optional
 from typing import Tuple
 from typing import Type
@@ -94,3 +97,29 @@ def try_import() -> _DeferredImportExceptionContextManager:
 
     """
     return _DeferredImportExceptionContextManager()
+
+
+class _LazyImport(types.ModuleType):
+    """Module wrapper for lazy import.
+
+    This class wraps specified module and lazily import it when they are actually accessed.
+    Otherwise, `import optuna` becomes slower because it imports all submodules and
+    their dependencies (e.g., bokeh) all at once.
+    Within this project's usage, importlib override this module's attribute on the first
+    access and the imported submodule is directly accessed from the second access.
+
+    Args:
+        name: Name of module to apply lazy import.
+    """
+
+    def __init__(self, name: str) -> None:
+        super().__init__(name)
+        self._name = name
+
+    def _load(self) -> types.ModuleType:
+        module = importlib.import_module(self._name)
+        self.__dict__.update(module.__dict__)
+        return module
+
+    def __getattr__(self, item: str) -> Any:
+        return getattr(self._load(), item)
