+++ optuna/cli.py
@@ -31,12 +31,14 @@ import yaml
 
 import optuna
 from optuna._deprecated import deprecated
+from optuna._imports import _LazyImport
 from optuna.exceptions import CLIUsageError
 from optuna.exceptions import ExperimentalWarning
 from optuna.storages import RDBStorage
-from optuna.study import _dataframe
 from optuna.trial import TrialState
 
+
+_dataframe = _LazyImport("optuna.study._dataframe")
 
 _DATETIME_FORMAT = "%Y-%m-%d %H:%M:%S"
