+++ optuna/multi_objective/__init__.py
@@ -1,6 +1,9 @@
+from optuna._imports import _LazyImport
 from optuna.multi_objective import samplers  # NOQA
 from optuna.multi_objective import study  # NOQA
 from optuna.multi_objective import trial  # NOQA
-from optuna.multi_objective import visualization  # NOQA
 from optuna.multi_objective.study import create_study  # NOQA
 from optuna.multi_objective.study import load_study  # NOQA
+
+
+visualization = _LazyImport("optuna.multi_objective.visualization")
